Criado entrypoint na HTTPAPI
-- Implementado entrypoint para pegar as ordens em aberto
-- Implementado função de fetch na store orderbook
